- Added image and font resources and included sample config for creating overlay.-Separated client library for future module creation.
