+++ .gitignore
@@ -1,1 +1,2 @@
-EyesonAPI/__pycache__/eyeson.cpython-311.pyc
+eyeson/__pycache__/eyeson.cpython-311.pyc
+**/.DS_Store
