+++ scripts/create_image_overlay.py
