+++ setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup
+
+setup(
+    name='eyeson',
+    version='0.0.1',
+    packages=['eyeson'],
+    url='https://www.eyeson.com/',
+    license='GNU v3',
+    author='Dinesh Shenoy',
+    author_email='dinesh.shenoy@eyeson.com',
+    description='Eyeson API Python SDK'
+)
